Fixed that when adding a new ArticleSummaryExtractorConfig gets added no delete button is shown

+++ DeepThoughtAndroidApp/src/main/kotlin/net/dankito/deepthought/android/dialogs/ArticleSummaryExtractorConfigDialog.kt
@@ -69,7 +69,7 @@ class ArticleSummaryExtractorConfigDialog {
             saveConfig(config, dialog, activity, callback)
         }
 
-        if(config.extractor is IImplementedArticleSummaryExtractor == false) {
+        if(config.extractor is IImplementedArticleSummaryExtractor == false && config.isPersisted() == true) {
             builder.setNeutralButton(R.string.action_delete) { dialog, _ ->
                 deleteConfig(config, dialog, activity)
             }
